refactor: move cmd/collector/app/sampling/model types to storage
