+++ plugin/sampling/strategyprovider/adaptive/provider_test.go
@@ -14,10 +14,10 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
 
-	"github.com/jaegertracing/jaeger/cmd/collector/app/sampling/model"
 	epmocks "github.com/jaegertracing/jaeger/plugin/sampling/leaderelection/mocks"
 	"github.com/jaegertracing/jaeger/proto-gen/api_v2"
 	smocks "github.com/jaegertracing/jaeger/storage/samplingstore/mocks"
+	"github.com/jaegertracing/jaeger/storage/samplingstore/model"
 )
 
 func TestProviderLoadProbabilities(t *testing.T) {
