+++ plugin/storage/badger/samplingstore/storage_test.go
@@ -12,8 +12,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	samplemodel "github.com/jaegertracing/jaeger/cmd/collector/app/sampling/model"
 	"github.com/jaegertracing/jaeger/pkg/testutils"
+	samplemodel "github.com/jaegertracing/jaeger/storage/samplingstore/model"
 )
 
 func newTestSamplingStore(db *badger.DB) *SamplingStore {
