+++ plugin/storage/es/samplingstore/dbmodel/converter_test.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/jaegertracing/jaeger/cmd/collector/app/sampling/model"
 	"github.com/jaegertracing/jaeger/pkg/testutils"
+	"github.com/jaegertracing/jaeger/storage/samplingstore/model"
 )
 
 func TestConvertDependencies(t *testing.T) {
